Improve searching lost measurements.
Very usual users have a situation where they move a pattern and measurements and place them in one folder. If path to measurements is not correct anymore but in the pattern folder we can find measurements file with the same name as expected in the broken path we will propose it by default. This change brings another case. Telegram automatically replace all spaces with underscores. Such names also must be checked and proposed.