Log pattern tool used.
It would be interesting to see popular and less popular tools, and to, at some point, take a deeper dive as to why people use the less often used tools, less often. It brings many interesting discussions - perhaps a tool doesn't quite work as intended or could be improved on.