Refactoring.
Variables should be initialized before use.
Types and variables should be declared in separate statements.
The "_t" and "_v" version of type traits should be used instead of "::type" and "::value".
String literals with different prefixes should not be concatenated.
Objects should not be sliced.
Non-reentrant POSIX functions should be replaced with their reentrant versions.
Inherited functions should not be hidden.
Empty statements should be removed.
Default capture should not be used.
"static" should not be used in unnamed namespaces.
"static" members should be accessed statically.
"empty()" should be used to test for emptiness.
Nested blocks of code should not be left empty.
Move and swap operations should be "noexcept".
Mergeable "if" statements should be combined.